SKALE-2123 libskutils refactored for further changes
