+++ libskutils/CMakeLists.txt
@@ -5,13 +5,14 @@ add_library( skutils STATIC
     ./src/command_line_parser.cpp   ./include/skutils/command_line_parser.h
     ./src/console_colors.cpp        ./include/skutils/console_colors.h
     ./src/dispatch.cpp              ./include/skutils/dispatch.h
-	./src/eth_utils.cpp             ./include/skutils/eth_utils.h
+    ./src/eth_utils.cpp             ./include/skutils/eth_utils.h
+    ./src/http.cpp                  ./include/skutils/http.h
     ./src/mail.cpp                  ./include/skutils/mail.h
                                     ./include/skutils/multifunction.h
     ./src/multithreading.cpp        ./include/skutils/multithreading.h
     ./src/network.cpp               ./include/skutils/network.h
-	./src/rest_call.cpp             ./include/skutils/rest_call.h
-	./src/stats.cpp                 ./include/skutils/stats.h
+    ./src/rest_call.cpp             ./include/skutils/rest_call.h
+    ./src/stats.cpp                 ./include/skutils/stats.h
     ./src/thread_pool.cpp           ./include/skutils/thread_pool.h
     ./src/url.cpp                   ./include/skutils/url.h
     ./src/utils.cpp                 ./include/skutils/utils.h
@@ -19,13 +20,14 @@ add_library( skutils STATIC
     ./src/btrfs.c                   ./include/skutils/btrfs.h
     )
 if( APPLE )
-	target_compile_definitions( skutils PRIVATE __BUILDING_4_MAC_OS_X__=1 )
+    target_compile_definitions( skutils PRIVATE __BUILDING_4_MAC_OS_X__=1 )
 endif()
 target_include_directories( skutils PRIVATE
     ./include
-	${SKUTILS_INCLUDE_DIRS}
-	"${DEPS_INSTALL_ROOT}/include" "${DEPS_INSTALL_ROOT}/include/cryptopp"
+    ${SKUTILS_INCLUDE_DIRS}
+    "${DEPS_INSTALL_ROOT}/include" "${DEPS_INSTALL_ROOT}/include/cryptopp"
     )
 target_link_libraries( skutils
     websockets uv pthread
     )
+
